Add basic google analytics event tracking
