+++ src/index.js
@@ -2,11 +2,5 @@ import App from "./App";
 import React from "react";
 import ReactDOM from "react-dom";
 
-import { BrowserRouter as Router } from "react-router-dom";
 const rootElement = document.getElementById("root");
-ReactDOM.render(
-  <Router>
-    <App />{" "}
-  </Router>,
-  rootElement
-);
+ReactDOM.render(<App />, rootElement);
